UpLoad Desafio 6

+++ src/main/java/trilha/back/financys/service/CategoriaService.java
@@ -1,9 +1,10 @@
 package trilha.back.financys.service;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
+import trilha.back.financys.dto.CategoriaDTO;
 import trilha.back.financys.entity.Categoria;
 import trilha.back.financys.repository.CategoriaRepository;
 
@@ -11,23 +12,50 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class CategoriaService {
+public class CategoriaService{
 
     @Autowired
     private CategoriaRepository categoriaRepository;
 
+
+
+
     public Categoria salvar(Categoria categoria){
-        return categoriaRepository.save(categoria);
+            return categoriaRepository.save(categoria);
     }
 
-    public List<Categoria> listaCategoria(){
-        return categoriaRepository.findAll();
+    public List<CategoriaDTO> listaCategoria(){
+        List<Categoria> listaCategoria = categoriaRepository.findAll();
+        return CategoriaDTO.convert(listaCategoria);
     }
 
     public Optional<Categoria> buscarPorId(Long id){
         return categoriaRepository.findById(id);}
 
-    public void removerPorId(Long id){
-        categoriaRepository.deleteById(id);
+    public String removerPorId(Long id){
+        String ret = "";
+        Optional<Categoria> cat = buscarPorId(id);
+        if (cat.isPresent()) {
+            categoriaRepository.deleteById(id);
+            ret = "Categoria deletada com sucesso!";
+        }
+        else
+            ret = "Categoria nao encontrada!";
+
+        return ret;
+    }
+
+    public String atualizarPorId(Long id, Categoria categoria) {
+        Optional<Categoria> cat = buscarPorId(id);
+        String ret = "";
+        if (cat.isPresent()) {
+           salvar(categoria);
+            ret = "Categoria Atualizada com sucesso!";
+        } else {
+            ret = "Categoria nao encontrada!";
+        }
+
+        return ret;
     }
 }
+
